+++ client/src/components/rapid/Line.js
@@ -1,43 +1,43 @@
 import React from "react";
-import PropTypes from "prop-types";
+import { buildPath, useDrag, useMousePosition } from "../shared/_utils";
+import Point from "../shared/Point";
+import Cubic from "../shared/Cubic";
 
 /**
  * A line represents a safe route to navigate the river at a specific water level
  */
 
-const Line = ({ name, desc, vector, x, y }) => {
-  const vectorArray = vector.split(/[\sa-zA-Z]+/).filter((elem) => elem);
-  let circleX = vectorArray[0];
-  let circleY = vectorArray[1];
-
-  return (
-    <g className="line" transform={`translate(${x},${y})`}>
-      <path
-        d={vector}
-        data-tip={`<div>
-            <div class="name">${name}</div>
-            <div class="desc">${desc}</div>
-          </div>`}
-        data-html={true}
-        data-for="svgTooltip"
-        data-event="click"
-      />
-      <circle cx={circleX} cy={circleY} r="10" stroke="none" fill="black" />
-    </g>
-  );
-};
-
-Line.propTypes = {
-  /** Name of the line*/
-  name: PropTypes.string.isRequired,
-  /** Description of the line. */
-  desc: PropTypes.string,
-  /** path vector of the line. */
-  vector: PropTypes.string.isRequired,
-  /** x coordinate of the line. */
-  x: PropTypes.number.isRequired,
-  /** y coordinate of the line. */
-  y: PropTypes.number.isRequired,
-};
+const Line = ({ line }) => (
+  <>
+    <path
+      className="ad-Path"
+      d={buildPath({ points: line, closePath: false })}
+    />
+    {line.map((p, i, a) => {
+      let anchors = [];
+      if (p.c) {
+        anchors.push(
+          <Cubic
+            index={i}
+            p1x={a[i - 1].x}
+            p1y={a[i - 1].y}
+            p2x={p.x}
+            p2y={p.y}
+            x1={p.c[0].x}
+            y1={p.c[0].y}
+            x2={p.c[1].x}
+            y2={p.c[1].y}
+          />
+        );
+      }
+      return (
+        <>
+          <Point index={i} x={p.x} y={p.y} />
+          {anchors}
+        </>
+      );
+    })}
+  </>
+);
 
 export default Line;
