+++ client/src/components/shared/SVG.js
@@ -1,0 +1,60 @@
+import React from "react";
+import { Table } from "react-bootstrap";
+import Line from "../rapid/Line";
+import { useSelector } from "react-redux";
+import Point from "../shared/Point";
+import Cubic from "../shared/Cubic";
+import { buildPath, useDrag, useMousePosition } from "../shared/_utils";
+import store from "../../rematch/store";
+import { useRef, useEffect } from "react";
+import PropTypes from "prop-types";
+import ReactTooltip from "react-tooltip";
+import Eddy from "./Eddy";
+import Hydraulic from "./Hydraulic";
+import Symbol from "./Symbol";
+import Popover from "./Popover";
+import filterRange from "./_utils/filterRange";
+
+const SVG = ({ level, rapid }) => {
+    
+
+    const mounted = useRef();
+
+    useEffect(() => {
+        if (!mounted.current) {
+          mounted.current = true;
+        } else {
+          ReactTooltip.rebuild();
+        }
+      });
+  
+    const coords = useMousePosition();
+  
+    const handleMouseMove = (e) => {
+      if (draggedPoint) {
+        dispatch.testEnvironment.setPointCoords({ coords });
+      } else if (draggedCubic !== false) {
+        dispatch.testEnvironment.setCubicCoords({ coords, anchor: draggedCubic });
+      }
+    };
+  
+    return (
+      <svg
+        className="Features"
+        id="vector-container"
+        viewBox="0 0 100 100"
+        xmlns="http://www.w3.org/2000/svg"
+        preserveAspectRatio="none"
+        onMouseMove={(e) => handleMouseMove(e)}
+        onMouseUp={() => dispatch.testEnvironment.cancelDragging()}
+      >
+        <g>
+          {lines.map((line) => (
+            <Line line={line} />
+          ))}
+        </g>
+      </svg>
+    );
+}
+
+export default SVG
