+++ client/src/components/shared/_utils/index.js
@@ -1,14 +1,10 @@
-import useDrag from "./useDrag";
-import usePathParser from "./usePathParser";
 import useWindowDimensions from "./useWindowDimensions";
 import PositiveNumber from "./positiveNumber";
 import buildPath from "./buildPath";
 import useMousePosition from "./useMousePosition";
 
 export {
-  useDrag,
   useWindowDimensions,
-  usePathParser,
   PositiveNumber,
   buildPath,
   useMousePosition,
